requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
segmentation-models-pytorch==0.3.3
tifffile==2023.7.18
ultralytics
opencv-python-headless==4.8.0.76
matplotlib==3.7.2
